# TBS: perf: Reduce allocs slightly in write and delete